Merge of query history extension PR
* Presto server extension

* Provide an extension API for customized functionality.

* Presto query history extension

The extension (PrestoQueryHistoryStore) will save a query into some database, when it's expired (completed) from the execution cache.
We can later get the query back by using query id (e.g. via REST api) to see the detailed information.

* Test naming pattern

This test was ignored without the "Test" prefix.

* Test QueryInfo serde to/from json

The test is to serialize a QueryInfo to json, then deserialize it
to a QueryInfo.

* Read query_info as a Clob

As it is a LONGTEXT column, we need to read it as a Clob.

* Test on QueryHistorySQLStore

Test saving and reading query history.
Also, fix some mistakes regarding to using jdbc API (e.g. close connection).

* Fix some check style error

* Separate extension implementation from presto-main

The extension implementation should be a plugin that will be dropped in when necessary.

* [Fix build] Use violation file from root

When we create the new module presto-history-extension, we forgot to define the violation file location.

* Add a DAO layer for db manipulation

* Use jdbi to reduce the boiler plate code of jdbc.

* Simplify the history extension config

All configurations (including the implementation class and jdbc connection configurations) are in one same place (a properties file).
Rely on a jvm property (extension.query.history.confi) to find the properties file and rely on the file to create the extension, disable extension if there is a failure during the process or the property is not defined.

* Provide a readme to explain the usage.

* Move query_history DDL to DAO

* Use presto.env and presto.dc from properties file

Or use presto.cluster to define the presto cluster.

* Update extension readme

* Get cluster from properties

Remove the hard-coded cluster, and use the config to discover it.

* Provide ser/deser methods

Do not expose ObjectMapper to public, but provide ser/deser API instead to avoid leak of abstraction.

* Checkstyle fix: remove a blank line

* Test with MariaDB

Use MariaDB4J to create integration tests.

* Formatize sql statements

Use capital sql keywords.

* Provide ddl as a method in QueryHistoryDAO

The QueryHistorySQLStore can thus try to create the table during initialization, if the table does not exist yet.

* Remove unused test

* Use "presto.cluster" to define cluster in the extension config

If this property is not defined, an error will be thrown during QueryHistorySQLStore initialization and the extension will not work.

* Provide a sample for properties file

Always try to look up extension config in etc/extension/query-history-store.properties

* Throw NoSuchElementException if cannot find query from history store

This exception is expected by the REST API.

* SqlPath equals: compare rawPath

rawPath is the source of truth of the SqlPath, while parsedPath could be uninitialized when calling equals.

* Json serialization: ignore SqlPath.parsedPath

For serialization, it only matters to serialize rawPath which is the source of truth. parsedPath is generated from rawPath, and it was preventing us from serializing a SqlPath with empty rawPath.

* Use mysql driver

Mysql driver is already bundled with presto jars, so we could just reuse it instead of having to add a dependency on mariadb-java-connector..

* Remove query-history-store.properties

For users who need it, just recreate it by following presto-history-extension/README.md

* Add CREATE_TABLE_TEST query

The CREATE_TABLE (with column compression) is not yet supported in a test environment. So we use CREATE_TABLE_TEST query to create test tables.
Also, we need to create table in prod/preprod before enabling this extension.

* Remove the commented line in the README

* Package presto-history-extension into presto-server

* Build presto-history-extension before presto-server

So that presto-server packaging could use presto-history-extension jar.

* Add mysql-connector to presto-server classpath

Package mysql-connector and presto-history-extension as a zip, and unzip it in presto-server lib

* Cosmetic

Fix a typo in README
Fix a log line for QueryHistorySQLStore